feat: <:loading /> slot for server-rendered markup
Adds a new slot called `<:loading />` that allows a user to server
render markup to display while waiting for the Svelte component to mount
on the client.

This is useful if you want to display default markup to prevent the page
from jumping once the client mounts, or to indicate to the user that
something is loading.

- [ ] Integrate with front-end and ensure it is removed after loading
- and through updates
- [ ] Consider raising if you have `ssr: true`
- [ ] Update documentation
- [ ] Add tests

Closes #166